fitness-ai-app/backend/requirements.txt
Django
djangorestframework
redis
django-redis
anthropic
python-dotenv
gunicorn
django-cors-headers
celery
Pillow
cryptography